convis-api/requirements.txt
fastapi>=0.115.2
uvicorn>=0.31.1
gunicorn>=21.2.0
pymongo==4.5.0
python-dotenv>=1.1.0
bcrypt==4.1.1
pydantic>=2.11.0
pydantic-settings>=2.5.2
python-multipart>=0.0.18
email-validator==2.3.0
anyio>=4.7.0
PyJWT==2.8.0
streamlit>=1.29.0
requests>=2.31.0
websockets>=12.0
twilio>=8.10.0
# Document processing
PyPDF2>=3.0.0
openpyxl>=3.1.0
python-docx>=1.1.0
# Enhanced PDF processing with OCR support
pymupdf>=1.23.0
pytesseract>=0.3.10
Pillow>=10.0.0
# Vector embeddings and search
openai>=1.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
# Vector database for fast similarity search
chromadb>=0.4.0
sentence-transformers>=2.2.0
